fixed issue caused by loader-spinner and added automatic refresh
