+++ client/src/screens/product.js
@@ -1,59 +1,30 @@
 import React, {useEffect, useState} from 'react';
 import { useParams } from 'react-router-dom';
-import Loader from './loader';
-import ReviewCard from './review_card';
+import ReviewCard from './reviewCard';
 import CollectionsPopup from './collectionsPopup';
+import ReviewPopup from './reviewPopup';
 
 const Product = () => {
-    const [productInfo, setProductInfo] = useState({})
     const [isDataLoading, setIsDataLoading] = useState(false);
-    const [filterDisplay, setFilterDisplay] = useState(false);
-    const [addReviewDisplay, setAddReviewDisplay] = useState(false)
-    const [newReview, setNewReview] = useState({username: "test-acc", rating: "", review: ""});
-    const [reviews, setReviews] = useState([]);
+
+    const [productInfo, setProductInfo] = useState({})
+    
     const [showCollectionsPopup, setShowCollectionsPopup] = useState(false);
+    const [addReviewPopup, setAddReviewPopup] = useState(false)
+
+    const [filterDisplay, setFilterDisplay] = useState(false);
+    
+    /* const [newReview, setNewReview] = useState({username: "test-acc", rating: "", review: ""}); */
+    const [reviews, setReviews] = useState([]);
     const { productID } = useParams()
 
     const filterPopup = () => {
         setFilterDisplay(!filterDisplay)
     }
 
-
-    const addReviewPopup = () => {
-        setAddReviewDisplay(!addReviewDisplay)
-    }
-
-    const handleReviewChange = (e) => {
-        setNewReview({...newReview, [e.target.name]: e.target.value})
-    };
-
     const handleNewComment = () => {
         loadReviews()
     }
-
-    const submitReview = async (e) => {
-        e.preventDefault();
-        try {
-            const response = await fetch(`/api/v1/reviews?productID=${productID}`, {
-                method: 'POST',
-                headers: {
-                    'Content-Type': 'application/json',
-                },
-                body: JSON.stringify(newReview)
-            });
-    
-            if (response.ok) {
-                console.log('review uploaded successfully');
-                setNewReview({ username: 'test-acc', rating: '', review: ''}); // Reset form
-                setAddReviewDisplay(false);
-                loadReviews();
-            } else {
-                console.error('Failed to submit review');
-            }
-        } catch (error) {
-            console.error('Error:', error);
-        }
-    };  
 
     const loadProductInfo = async () => {
         //const apiUrl = process.env.REACT_APP_API_URL; // Use the API URL from environment variables
@@ -71,29 +42,14 @@ const Product = () => {
     }  
 
     const loadReviews =  async () => {
-        setIsDataLoading(true);
         await fetch(`/api/v1/reviews?productID=${productID}`)
             .then((res) => res.json())
             .then((data) => {
                 console.log(data);
                 setReviews(data);
-                setIsDataLoading(false);
             })
             .catch((error) => console.error('Error loading reviews:', error))
     }
-
-    /* const loadCollections =  async () => {
-        await fetch(`/api/v1/collections`)
-            .then((res) => res.json())
-            .then((data) => {
-                console.log("recieved collection data", data);
-                setCollections(data);
-                console.log("saved collections data", collections)
-                setCollectionsDisplay(true)
-            })
-            .catch((error) => console.error('Error loading collections:', error))
-    }
- */
     useEffect(() => {
         loadProductInfo()
     }, []);
@@ -118,8 +74,6 @@ const Product = () => {
     only be rendered once and then basically cached.
 
     POSTs:
-
-    - Add product to collection: user must pick either a default collection (wishlist) or another collection
     - Write a review: select existing tags, dropdown rating, review text
 
      */
@@ -129,7 +83,7 @@ const Product = () => {
     /* determined by user so implement persistence later */
     return (
         <div>
-            {isDataLoading ? <Loader /> : 
+            {isDataLoading ? <></> : 
             <>
             <div className="product-container">
 
@@ -161,44 +115,20 @@ const Product = () => {
                 </div>
 
                 <button onClick={() => setShowCollectionsPopup(!showCollectionsPopup)} id="add-to-collection">Add to Collection <span className="fa fa-angle-down down-arrow"></span></button>
+                
                 {showCollectionsPopup ? <CollectionsPopup productID={productID} callback={() => {setShowCollectionsPopup(!showCollectionsPopup)}}/> : <></>}
+                
                 <hr />
 
-                {addReviewDisplay ? 
-                <>
-                <div className="filter-overlay"></div>
-                <div className="add-review-popup">
-                    <div className="popup-head">
-                        <h4>Write a review</h4>
-                        <button onClick={addReviewPopup}><span className="fa fa-minus"></span></button>
-                    </div>
-                    <form id="add-review" onSubmit={submitReview}>
-                        <p>Rating</p>
-                        <select className="form-select" name="rating" value={newReview.rating} onChange={handleReviewChange} aria-label='Choose rating'>
-                            <option selected></option>
-                            <option value="1">1</option>
-                            <option value="2">2</option>
-                            <option value="3">3</option>
-                            <option value="4">4</option>
-                            <option value="5">5</option>
-                        </select>
-                        <p>What did you think about this product?</p>
-                        <textarea rows={4} name="review" value={newReview.review} onChange={handleReviewChange}/>
-                        <button type="submit">Submit</button>
-                    </form>
-                </div> 
-                </>
-                : 
-                <></>
-
-                }
+                {addReviewPopup ? <ReviewPopup productID={productID} callback={() => {setAddReviewPopup(!addReviewPopup)}} render={() => loadReviews()}/> : <></>}
+            
             </div>
             
         </div>
         <div id="product-reviews">
             <div className="product-reviews-head">
                 <h3>REVIEWS ({reviews.length})</h3>
-                <button onClick={addReviewPopup} id="add-review-button">Write a Review</button>
+                <button onClick={() => setAddReviewPopup(!addReviewPopup)} id="add-review-button">Write a Review</button>
                 <button onClick={filterPopup} id="sort-reviews">Sort</button>
                 {filterDisplay ? 
                 <>
@@ -222,7 +152,7 @@ const Product = () => {
             </div>
             {/* reviews should be handled in a separate component */}
             <div id='reviews'>
-                {isDataLoading ? <Loader /> : reviews.map((review) => <ReviewCard key={review._id} review={review} callback={handleNewComment}/>)}     
+                {isDataLoading ? <></> : reviews.map((review) => <ReviewCard key={review._id} review={review} callback={handleNewComment}/>)}     
             </div>
         </div>
     </>
